individual chart widget sharing
