+++ src/ducks/Studio2/Widget/creators.js
@@ -1,0 +1,2 @@
+export { newChartWidget } from './ChartWidget'
+export { newHolderDistributionWidget } from './HolderDistributionWidget'
